requirements.txt
# Core API
fastapi>=0.110,<0.112
uvicorn[standard]>=0.29,<0.31

# DB & ORM
sqlalchemy>=2.0,<2.1
alembic>=1.13,<1.14
psycopg[binary]>=3.1,<3.2

# Validation & settings
pydantic>=2.7,<2.9
pydantic-settings>=2.2,<2.4
email-validator>=2.1,<2.3

# Auth & Security
passlib[bcrypt]>=1.7,<1.8
python-jose[cryptography]>=3.3,<3.4

# Uploads & parsing
python-multipart>=0.0.9,<0.0.10

# Performance / qualidade (opcional, mas recomendado)
orjson>=3.10,<3.11

# Env helpers
python-dotenv>=1.0,<1.1

# Logging (opcional)
loguru>=0.7,<0.8
